notification-service/requirements.txt
django
djangorestframework
requests
PyJWT
django-cors-headers
dj-database-url
gunicorn
